feat: Add deletion of chosen domain data
Effect-UI: external
Effect-Dev: external

+++ src/entrypoints/sidepanel/components/DomainDataView.tsx
@@ -1,5 +1,8 @@
 import { commands, DomainData, URLMetaData } from "../commands";
 import { ConsoleContext } from "../contexts/ConsoleContext";
+import { handleResponse } from "../helpers";
+import { useCheckedMap } from "../hooks";
+import ButtonRow from "./ButtonRow";
 
 interface CombinedMetaData extends URLMetaData {
   url: string;
@@ -13,12 +16,14 @@ function useURLMetaData(): {
   const [domainData, setDomainData] = useState<DomainData>({ loopingData: {} });
   const { logger } = useContext(ConsoleContext);
 
+  // make a list of the data, flattening the url to be with the data
   const data = useMemo<CombinedMetaData[]>(() => {
     return Object.entries(domainData.loopingData).map(([url, dat]) => {
       return { ...dat, url };
     });
   }, [domainData]);
 
+  // create a set out of the tags
   const tagSet = useMemo<Set<string>>(() => {
     const allTags = Object.entries(data).flatMap(([_, { tags }]) => {
       if (!tags) return [];
@@ -28,6 +33,9 @@ function useURLMetaData(): {
     return new Set(allTags);
   }, [data]);
 
+  /**
+   * Update the domain data.
+   */
   async function update() {
     const response = await commands.loadDomainData();
     if (!response.success) {
@@ -48,7 +56,8 @@ function useURLMetaData(): {
 
 export default function DomainDataView() {
   const [textFilter, setTextFilter] = useState<string>("");
-  const [checked, setChecked] = useState<Record<string, unknown>>({});
+  const tagMap = useCheckedMap<string>({ defaultValue: null });
+  const urlDataMap = useCheckedMap<string>({ defaultValue: null });
   const urlData = useURLMetaData();
   const { logger } = useContext(ConsoleContext);
 
@@ -84,17 +93,7 @@ export default function DomainDataView() {
                   <input
                     type="checkbox"
                     name={tag}
-                    onChange={(ev) =>
-                      setChecked((prev) => {
-                        const ob = { ...prev };
-                        if (tag in ob) {
-                          delete ob[tag];
-                        } else {
-                          ob[tag] = "";
-                        }
-                        return ob;
-                      })
-                    }
+                    onChange={(ev) => tagMap.toggle(tag)}
                   />
                 </label>
               </li>
@@ -104,22 +103,34 @@ export default function DomainDataView() {
       </details>
       <ul>
         {urlData.data.map(({ url, ...val }) => {
+          /**
+           * Whether this entry has each of the chosen tags.
+           */
           const correctTags =
-            Object.keys(checked).length === 0 ||
+            tagMap.size === 0 ||
             !!val.tags?.reduce((prev, cur) => {
-              return prev || Object.keys(checked).includes(cur);
+              return prev || tagMap.has(cur);
             }, false);
 
+          /**
+           * Whether this entry matches the text filter.
+           */
           const matchesTextFilter =
             textFilter === "" ||
             !!val.title?.toLowerCase().includes(textFilter) ||
-            val.description?.toLowerCase().includes(textFilter);
+            !!val.description?.toLowerCase().includes(textFilter);
 
           if (!matchesTextFilter || !correctTags) return null;
 
           return (
             <li key={url}>
               <a href={url}>{`${val.title ?? "<No title>"} (${url})`}</a>
+              <input
+                type="checkbox"
+                onChange={(ev) => {
+                  urlDataMap.toggle(url);
+                }}
+              />
               <p>{val.description ?? ""}</p>
               <ul>
                 {val.tags?.map((tag) => (
@@ -130,15 +141,39 @@ export default function DomainDataView() {
           );
         })}
       </ul>
-      <button
-        type="button"
-        onClick={async () => {
-          await commands.deleteDomainData();
-          urlData.update();
-        }}
-      >
-        Delete domain data
-      </button>
+      <ButtonRow>
+        <button
+          type="button"
+          onClick={async () => {
+            if (window.confirm("Delete all domain data?")) {
+              await commands.deleteDomainData();
+              urlData.update();
+            }
+          }}
+        >
+          Delete all
+        </button>
+        <button
+          type="button"
+          disabled={urlDataMap.size < 1}
+          onClick={async () => {
+            if (
+              urlDataMap.size > 0 &&
+              window.confirm("Delete chosen domain data?")
+            ) {
+              const response = await commands.deleteDomainUrlData({
+                urls: [...urlDataMap.keys()],
+              });
+              handleResponse(response, logger);
+              // assume success
+              urlDataMap.clear();
+              urlData.update();
+            }
+          }}
+        >
+          Delete chosen
+        </button>
+      </ButtonRow>
     </div>
   );
 }
